Rename to redirect.tsx
